FieldMatches.set_orphans_to_combed_matches: Use SSIM to figure out similarity between next/prev frame
This should help make smarter decisions about when to change a match to a 'c' match (which would in turn be a match that must be deinterlaced). This is not foolproof, however; Some minor animation changes, such as smoke or very small mouth flaps, will be missed this way. However, this is still overall much safer, and catches at least the most obvious large-frame differences or similarities relatively reliably.